+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/BoxTestGenerated.java
@@ -1640,6 +1640,12 @@ public class BoxTestGenerated extends AbstractBoxTest {
       }
 
       @Test
+      @TestMetadata("GraphPrivateMultibindingsAreGraphLocal.kt")
+      public void testGraphPrivateMultibindingsAreGraphLocal() {
+        run("GraphPrivateMultibindingsAreGraphLocal.kt");
+      }
+
+      @Test
       @TestMetadata("GraphPrivatePreventsLeakingParentScopedInstance.kt")
       public void testGraphPrivatePreventsLeakingParentScopedInstance() {
         run("GraphPrivatePreventsLeakingParentScopedInstance.kt");