+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/JsFastInitBoxTestGenerated.java
@@ -1378,6 +1378,12 @@ public class JsFastInitBoxTestGenerated extends AbstractJsFastInitBoxTest {
       }
 
       @Test
+      @TestMetadata("GraphPrivateMultibindingsAreGraphLocal.kt")
+      public void testGraphPrivateMultibindingsAreGraphLocal() {
+        run("GraphPrivateMultibindingsAreGraphLocal.kt");
+      }
+
+      @Test
       @TestMetadata("GraphPrivatePreventsLeakingParentScopedInstance.kt")
       public void testGraphPrivatePreventsLeakingParentScopedInstance() {
         run("GraphPrivatePreventsLeakingParentScopedInstance.kt");